pyproject.toml
[project]
name = "texteller4chemical"
version = "0.1.0"
description = "Fine-tune pretrained TexTeller on EDU-CHEMC handwritten chemical structure recognition."
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "accelerate>=0.29.0",
    "datasets>=2.19.0",
    "evaluate>=0.4.2",
    "numpy>=1.26.0",
    "opencv-python>=4.9.0.80",
    "pandas>=2.2.0",
    "peft>=0.10.0",
    "pillow>=10.0.0",
    "pyyaml>=6.0.1",
    "rapidfuzz>=3.8.0",
    "tensorboard>=2.16.0",
    "torch>=2.11.0",
    "torchvision>=0.26.0",
    "tqdm>=4.66.0",
    "transformers>=4.40.0",
]

[project.optional-dependencies]
texteller = [
    "texteller",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/chemtexteller"]

[tool.uv.sources]
torch = { index = "pytorch-cu130" }
torchvision = { index = "pytorch-cu130" }

[[tool.uv.index]]
name = "pytorch-cu130"
url = "https://download.pytorch.org/whl/cu130"
explicit = true
